adding draw rectangle
I've added two functions for drawing a rectangle:
RectFromUpperLeft(...) and RectFromLowerLeft(..)
but I think only one of them should go to the master and with a simpler name (for example just Rect() or Rectangle() )
Two because it's common in gopdf to draw from upper-left corner, but in PDF specification rectangle goes from lower-left. I don't know what is better here.