Introduce SplitTextWithOption to allow configurable text splitting.
To have a higher level of flexibiliity and reduce repetitive code SplitTextWithOption is now the underlying text splitting implementation for SplitText and SplitTextWithWordWrap. Splitting can now be performed individually based on language or font specifics. Two break modes are now available to process texts either with a strict break (allowing mid-word breaks) or with an identifier-sensitive mode (avoiding mid-word breaks). See the code documentation of BreakOption for more info.