Fix go vet failure
When running the go vet linter on the entire project, we encounter a failure
stating that an integer to string conversion using "string(c)" can lead to
unexpected results. Such a failure might become an error in future
versions of Go. See for example https://go.dev/doc/go1.15#vet.

By explicitly converting the integer to a rune before casting it to a
string, we avoid this issue and can run go test without the "-vet=off"
flag.